% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bcse}
\alias{bcse}
\title{16S amplicon dataset from the GLBRC Biofuel Cropping System Experiment
(BCSE) at Michigan State University, Kellogg Biological Station}
\format{
An object class "phyloseq".
}
\source{
Internal test dataset.
}
\usage{
data("bcse")
}
\description{
A phyloseq containing 2861 97\% sequence similarity OTUs
(Operational Taxonomic Units) across 50 samples. Samples are for the leaf
microbiome only.
}
\references{
Haan, N. L., Benucci, G. N. M., Fiser, C. M., Bonito, G., &
Landis, D. A. (2023). Contrasting effects of bioenergy crops on
biodiversity. Science Advances, 9(38), eadh7960.
\url{doi:10.1126/sciadv.adh7960}
}
\keyword{datasets}
