% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bean}
\alias{bean}
\title{16S amplicon dataset from common bean (Phaseolus vulgaris)}
\format{
An object class "phyloseq".
}
\source{
Internal test dataset.
}
\usage{
data("bean")
}
\description{
A phyloseq containing 24885 ASVs (Amplicon Sequence Variants)
across 30 samples.
}
\references{
Shade A, Stopnisek N (2019) Abundance-occupancy distributions to
prioritize plant core microbiome membership. Current Opinion in
Microbiology, 49:50-58 \url{doi:10.1016/j.mib.2019.09.008}
}
\keyword{datasets}
