% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brc_pcoa.R
\name{brc_pcoa}
\alias{brc_pcoa}
\title{Calculate Principal Coordinate Analysis (PCoA) using the weighted classical multidimensional scaling (wcmdscale) method.}
\usage{
brc_pcoa(
  asv_matrix,
  physeq,
  ncores = parallel::detectCores(),
  k = 2,
  eig = TRUE,
  add = FALSE,
  x.ret = FALSE
)
}
\arguments{
\item{asv_matrix}{A matrix, data frame, or distance object containing the ASV (Amplicon Sequence Variant) data
to be used for Principal Coordinate Analysis.}

\item{physeq}{A phyloseq object containing sample metadata that will be incorporated into the PCoA results.}

\item{ncores}{Integer specifying the number of CPU cores to use for parallel processing.
Default is all available cores as detected by parallel::detectCores().}

\item{k}{Integer specifying the number of dimensions (axes) to calculate in the PCoA. Default is 2.}

\item{eig}{Logical indicating whether eigenvalues should be returned. Default is TRUE.}

\item{add}{Logical indicating whether to add a constant to the non-diagonal dissimilarities such that all
eigenvalues are non-negative. Default is FALSE.}

\item{x.ret}{Logical indicating whether the rotated configuration should be returned. Default is FALSE.

#' @export}
}
\description{
Calculate Principal Coordinate Analysis (PCoA) using the weighted classical multidimensional scaling (wcmdscale) method.
}
