% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_helpers.R
\name{detect_hpc_type}
\alias{detect_hpc_type}
\title{Detect HPC environment type}
\usage{
detect_hpc_type()
}
\value{
Character string identifying the HPC environment type:
"slurm", "pbs", "sge", "lsf", "condor", "generic_hpc", or "local"
}
\description{
Detects if code is running in an HPC environment and identifies which scheduler.
}
\examples{
detect_hpc_type()
}
