% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variance_propagation.R
\name{.hill_number}
\alias{.hill_number}
\title{Helper function to calculate hill numbers}
\usage{
.hill_number(x, q)
}
\arguments{
\item{x}{Sample seqeunce counts}

\item{q}{Hill number order (q = 0 for richness, q = 1 for Shannon, q = 2 for Simpson)}
}
\description{
Helper function to calculate hill numbers
}
