% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_core.R
\name{filter_core}
\alias{filter_core}
\title{Select ASVs Based on Occurrence Threshold}
\usage{
filter_core(physeq, threshold = 0.6, as = "rows")
}
\arguments{
\item{physeq}{A phyloseq object.}

\item{threshold}{A numeric value between 0 and 1 specifying the occurrence threshold.
Default is 0.6 (60\% of samples).}

\item{as}{The matrix dimension that is desired for taxa. Must be
"rows" for rows and "columns" or "cols" for columns.}
}
\value{
A named list containing:
\itemize{
\item \code{physeq_high_occ}: Phyloseq object with high occurrence ASVs.
\item \code{physeq_low_occ}: Phyloseq object with low occurrence ASVs.
}
}
\description{
This function selects ASVs from a phyloseq object based on their occurrence frequency
across samples. It returns filtered phyloseq objects for high and low occurrence ASVs.
}
