% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_helpers.R
\name{get_available_cores}
\alias{get_available_cores}
\title{Get the number of available CPU cores based on environment}
\usage{
get_available_cores(default = 1)
}
\arguments{
\item{default}{Default number of cores to use if detection fails}
}
\value{
Integer number of cores available
}
\description{
Determines the appropriate number of cores to use based on whether
we're in an HPC environment or a local machine.
}
\examples{
get_available_cores()
get_available_cores(default = 2)

}
