% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{gg_ordi}
\alias{gg_ordi}
\title{Standard Ordination Plot with ggplot2}
\usage{
gg_ordi(.data, .color, ordi, .drop_na = NULL)
}
\arguments{
\item{.data}{A data frame containing ordination coordinates and additional metadata.}

\item{.color}{A column in \code{.data} to use for coloring points and ellipses. This should be a categorical variable.}

\item{ordi}{A character string specifying the ordination method. Options are "NMDS" and "PcoA".}

\item{.drop_na}{A column in \code{.data} used to filter out rows with missing values. Rows with \code{NA} in this column will be removed.}

\item{.shape}{A column in \code{.data} to use for shaping points. This is optional and can be \code{NULL}.}
}
\value{
A \code{ggplot} object representing the NMDS plot with points, ellipses, and reference lines.
}
\description{
This function creates a standardized ordination plot using \code{ggplot2}.
It includes points, ellipses, and reference lines for visualizing NMDS or PCoA results. The function allows
customization of color, shape, and handling of missing values.
}
\examples{
# Load example data
library(vegan)
data(dune, package = "vegan")

# Run NMDS
nmds_result <- metaMDS(dune, k = 2)
nmds_scores <- scores(nmds_result, display = "sites")

# Adding factor to data
dune$Management <- sample(c("A", "B"), nrow(dune), replace = TRUE)
nmds_data <- cbind(nmds_scores, dune)

# Create NMDS plot
gg_ordi(nmds_data, .color = Management, .drop_na = Management)
}
