% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_helpers.R
\name{is_hpc_environment}
\alias{is_hpc_environment}
\title{Check if running in an HPC environment}
\usage{
is_hpc_environment()
}
\value{
Logical indicating whether we're in any HPC environment
}
\description{
Check if running in an HPC environment
}
\examples{
is_hpc_environment()

}
