% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mimulus}
\alias{mimulus}
\title{16S amplicon dataset from yellow monkeyflower (Mimulus guttatus)}
\format{
An object class "phyloseq".
}
\source{
Internal test dataset.
}
\usage{
data("mimulus")
}
\description{
A phyloseq containing 11085 ASVs (Amplicon Sequence Variants)
across 39 samples.
}
\references{
Shade A, Stopnisek N (2019) Abundance-occupancy distributions to
prioritize plant core microbiome membership. Current Opinion in
Microbiology, 49:50-58 \url{https://doi.org/10.1016/j.mib.2019.09.008}
}
\keyword{datasets}
