% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_rarefy.R
\name{multi_rarefy}
\alias{multi_rarefy}
\title{Run multiple rarefaction for microbiome count tables}
\usage{
multi_rarefy(
  physeq,
  depth_level,
  num_iter = 100,
  .summarize = TRUE,
  threads = get_available_cores(),
  set_seed = NULL
)
}
\arguments{
\item{physeq}{A \code{phyloseq} object containing an OTU/ASV table.}

\item{depth_level}{An integer specifying the sequencing depth (number of
OTUs/ASVs) to which samples should be rarefied.}

\item{num_iter}{An integer specifying the number of iterations to perform
for rarefaction.}

\item{.summarize}{A logical indicating whether to summarize the results by
averaging across iterations (default = TRUE). If FALSE, the function will
return a named list of data frames (e.g. iter_1), one for each iteration, without
averaging.}

\item{threads}{Number of threads (default = \code{get_available_cores()}).}

\item{set_seed}{An optional integer to set the random seed for reproducibility (default = NULL).}
}
\value{
A data frame with taxa as rows and samples as columns. The values
represent the average sequence counts calculated across all iterations.
Samples with less than \code{depth_level} sequences are discarded.
}
\description{
This function performs multiple rarefaction on a \code{phyloseq} object by randomly
sub-sampling OTUs/ASVs within samples without replacement. The process is
repeated for a specified number of iterations, and the results are averaged.
Samples with fewer OTUs/ASVs than the specified \code{depth_level} are discarded.
}
\examples{
\donttest{
library(phyloseq)
library(BRCore)
data("bcse", package = "BRCore")

# Example rarefaction (single iteration, single core to keep examples fast)
otu_table_rare <- multi_rarefy(
  physeq = bcse,
  depth_level = 1000,
  num_iter = 100,
  .summarize = TRUE,
  threads = 2,
  set_seed = 7642
)

rowSums(otu_table_rare)
}

}
