% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_rarefy_new.R
\name{multi_rarefy_new}
\alias{multi_rarefy_new}
\title{Iterated rarefaction for phyloseq objects}
\usage{
multi_rarefy_new(physeq, depth, n_iter = 1, .array = TRUE)
}
\arguments{
\item{physeq}{A phyloseq object containing an otu_table.}

\item{depth}{Rarefaction depth.}

\item{n_iter}{Number of rarefaction iterations.}

\item{.array}{Logical; if TRUE return a 3D array, otherwise a list of data.frames.}
}
\value{
\itemize{
\item If n_iter = 1: a data.frame (samples x taxa)
\item If n_iter > 1 and .array = TRUE: a 3D array (samples x taxa x iterations)
\item If n_iter > 1 and .array = FALSE: a list of data.frames
}
}
\description{
Performs one or multiple rarefactions using vegan::rrarefy().
}
