% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelly_rarefy.R
\name{parallelly_rarefy}
\alias{parallelly_rarefy}
\title{Run multiple rarefaction for microbiome count tables}
\usage{
parallelly_rarefy(
  physeq,
  depth_level,
  num_iter = 100,
  threads = get_available_cores(),
  set_seed = NULL
)
}
\arguments{
\item{physeq}{A \code{phyloseq} object containing an OTU/ASV table.}

\item{depth_level}{An integer specifying the sequencing depth (number of
OTUs/ASVs) to which samples should be rarefied.}

\item{num_iter}{An integer specifying the number of iterations to perform
for rarefaction.}

\item{threads}{Number of threads (default = 4).}

\item{set_seed}{An optional integer to set the random seed for reproducibility (default = 123).}
}
\value{
A data frame with taxa as rows and samples as columns. The values
represent the average sequence counts calculated across all iterations.
Samples with less than \code{depth_level} sequences are discarded.
}
\description{
This function performs multiple rarefaction on a \code{phyloseq} object by randomly
sub-sampling OTUs/ASVs within samples without replacement. The process is
repeated for a specified number of iterations, and the results are averaged.
Samples with fewer OTUs/ASVs than the specified \code{depth_level} are discarded.
}
\examples{
data(GlobalPatterns, package = "phyloseq")
test_otutable_rare <- parallelly_rarefy(
  physeq = GlobalPatterns,
  depth_level = 500,
  num_iter = 10,
  threads = 2,
  set_seed = 123
)

head(test_otutable_rare)

}
