% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_helpers.R
\name{setup_parallel_backend}
\alias{setup_parallel_backend}
\title{Set up an appropriate parallel backend}
\usage{
setup_parallel_backend(default = 1)
}
\arguments{
\item{default}{Default number of cores to use if detection fails}
}
\value{
Either a cluster object (HPC) or integer number of cores (local)
}
\description{
Creates a parallel cluster if in an HPC environment or returns
the number of cores to use if on a local machine.
}
\examples{
setup_parallel_backend()
}
