% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{subset_fasta}
\alias{subset_fasta}
\title{Select a subset of sequences from a fasta file}
\usage{
subset_fasta(
  file = NULL,
  subset = NULL,
  out = paste(file, ".subset", sep = "")
)
}
\arguments{
\item{file}{Fasta file}

\item{subset}{Vector with (exact) names of the sequences to be retrieved.}

\item{out}{Path to output file. If absent, the '.mod' is added to the input file's name (path is thus conserved).}
}
\value{
Writes a fasta file with the selected sequences.
}
\description{
This function loads a fasta file, selects a subset on sequences based on the sequence's names and writes a new fsata file.
}
\author{
Guillem Salazar \href{mailto:salazar@icm.csic.es}{salazar@icm.csic.es}
}
\keyword{FastaUtils}
