% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{switchgrass_core}
\alias{switchgrass_core}
\title{Identified core microbiome members for the switchgrass dataset}
\format{
An object class "phyloseq".
}
\source{
Internal test dataset.
}
\usage{
data("switchgrass_core")
}
\description{
A phyloseq containing the identified core microbiome members
for the switchgrass dataset. This object is used for BRCore function
examples.
It was generated by running the following on BRCore 2.0.2 (2026-04-30):

\if{html}{\out{<div class="sourceCode R">}}\preformatted{switchgrass_core <- identify_core(
  physeq_obj     = switchgrass,
  priority_var   = "sampling_date",
  increase_value = 0.02,
  seed           = 092825
)
}\if{html}{\out{</div>}}
}
\references{
Shade A, Stopnisek N (2019) Abundance-occupancy distributions to
prioritize plant core microbiome membership. Current Opinion in
Microbiology, 49:50-58 \url{https://doi.org/10.1016/j.mib.2019.09.008}
}
\keyword{datasets}
