% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_phyloseq}
\alias{test_phyloseq}
\title{Phyloseq object dataset}
\format{
An object class "phyloseq".
}
\source{
Internal test data for BRCore package.
}
\usage{
data("test_phyloseq")
}
\description{
A 16S amplicon dataset containing 474 OTU (Operational Taxonomic Unit) across 622 samples.
}
\references{
To we add this later or change dataset...
}
\keyword{datasets}
