% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_phyloseq.R
\name{do_phyloseq}
\alias{do_phyloseq}
\title{Add a rarefied otu_table to a phyloseq object}
\usage{
do_phyloseq(physeq, otu_rare)
}
\arguments{
\item{physeq}{A \code{phyloseq} object in which you want to add the rarefied OTU/ASV table.}

\item{otu_rare}{A rarefied otu_table dataframe.}
}
\value{
A \code{phyloseq} object.
}
\description{
This function imports a rarefied otu_table (or normalized) in a \code{phyloseq} object \code{otu_table()} layer.
}
\examples{
\donttest{
library(phyloseq)
data(GlobalPatterns, package = "phyloseq")

# Perform multiple rarefaction
otu_table_rare <- 
    parallel_rarefy(
        physeq = GlobalPatterns,
        depth_level = 200,
        num_iter = 3,
        threads = 1,
        set_seed = 123
    )

# Check the rarefied data output
rowSums(otu_table_rare)

# Recreate the phyloseq object and check
rarefied_GlobalPatterns<- 
    do_phyloseq(physeq = GlobalPatterns, 
                otu_rare = otu_table_rare )
    
sample_sums(rarefied_GlobalPatterns)
}

}
