% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_phyloseq.R
\name{do_phyloseq}
\alias{do_phyloseq}
\title{Import a rarefied otu_table into a phyloseq object}
\usage{
do_phyloseq(physeq, otu_rare)
}
\arguments{
\item{physeq}{A \code{phyloseq} object in which I want to add the rarefied OTU/ASV table.}

\item{otu_rare}{A rarefied otu_table dataframe.}
}
\value{
A \code{phyloseq} object.
}
\description{
This function imports a rarefied otu_table (or normalized) in a \code{phyloseq} object \code{otu_table()} layer.
}
\examples{
\dontrun{
# Load example data
data(GlobalPatterns, package = "phyloseq")

# Perform multiple rarefaction
rarefied_data <- multi_rarefy(GlobalPatterns, depth_level = 1000, num_iter = 99)
head(rarefied_data)

rarefied_physeq <- do_phyloseq(physeq = GlobalPatterns, otu_rare=rarefied_data )
head(otu_table(rarefied_physeq))
}
}
