% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mimulus}
\alias{mimulus}
\title{16S amplicon dataset from yellow monkeyflower (Mimulus guttatus)}
\format{
An object class "phyloseq".
}
\source{
Internal test dataset.
}
\usage{
data("mimulus")
}
\description{
A phyloseq containing 11085 ASVs (Amplicon Sequence Variants)
across 39 samples.
}
\references{
Shade A, Stopnisek N. Abundance-occupancy distributions to prioritize
plant core microbiome membership. Curr Opin Microbiol. 2019 Jun;49:50-58.
doi: 10.1016/j.mib.2019.09.008. Epub 2019 Nov 10. PMID: 31715441.
}
\keyword{datasets}
