% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_rarefy.R
\name{multi_rarefy}
\alias{multi_rarefy}
\title{Perform Multiple Rarefaction on a Phyloseq Object}
\usage{
multi_rarefy(physeq, depth_level, num_iter)
}
\arguments{
\item{physeq}{A \code{phyloseq} object containing an OTU/ASV table.}

\item{depth_level}{An integer specifying the sequencing depth (number of OTUs/ASVs) to which samples should be rarefied.}

\item{num_iter}{An integer specifying the number of iterations to perform for rarefaction.}
}
\value{
A data frame with samples as rows and taxa as columns. The values represent the average abundance across all iterations. Samples with less than \code{depth_level} sequences will be discarded.
}
\description{
This function performs multiple rarefaction on a \code{phyloseq} object by randomly sub-sampling OTUs/ASVs within samples without replacement. The process is repeated for a specified number of iterations, and the results are averaged. Samples with fewer OTUs/ASVs than the specified \code{depth_level} are discarded.
}
\examples{
# Load example data
data(GlobalPatterns, package = "phyloseq")

# Perform multiple rarefaction
rarefied_data <- multi_rarefy(GlobalPatterns, depth_level = 1000, num_iter = 10)
head(rarefied_data)
}
