% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variance_propagation.R
\name{plot_variance_propagation}
\alias{plot_variance_propagation}
\title{Variance propagation diagnostic for rarefaction}
\usage{
plot_variance_propagation(physeq, rarefied, q = 0, group_var, group_color)
}
\arguments{
\item{physeq}{Raw phyloseq object}

\item{rarefied}{Output from iterated_rarefy_phyloseq()}

\item{q}{Hill number order (alpha only)}

\item{group_var}{grouping variable in sample_data}

\item{measure}{"alpha" or "beta"}

\item{distance}{"bray" or "jaccard" (beta only)}
}
\value{
ggplot
}
\description{
Compare raw vs rarefied diversity distributions across iterations.
}
