% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{switchgrass}
\alias{switchgrass}
\title{16S amplicon dataset from Switchgrass (Panicum virgatum)}
\format{
An object class "phyloseq".
}
\source{
Internal test dataset.
}
\usage{
data("switchgrass")
}
\description{
A phyloseq containing 706 OTUs (Operational Taxonomic Units) across 43 samples.
}
\references{
Shade A, Stopnisek N. Abundance-occupancy distributions to prioritize plant core microbiome membership. Curr Opin Microbiol. 2019 Jun;49:50-58. doi: 10.1016/j.mib.2019.09.008. Epub 2019 Nov 10. PMID: 31715441.
}
\keyword{datasets}
